Added simple command line interface for playing the game.
The command line interface allows two human players to play against each other, or for a human player to play against an automatic agent. The only agent implemented for the moment is a random player, which selects with uniform probability a move from all the valid choices at each turn.